requirements.txt
pylint
asgiref==3.8.1
astroid==3.3.7
colorama==0.4.6
diff-match-patch==20241021
dill==0.3.9
Django==5.1.3
django-filter==24.3
django-import-export==4.3.3
django-simple-history==3.7.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
isort==5.13.2
mccabe==0.7.0
sqlparse==0.5.2
tablib==3.7.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
